+++ app/client/src/IDE/Components/EditableName/index.ts
@@ -1,0 +1,1 @@
+export { EditableName } from "./EditableName";
