+++ app/client/src/pages/Editor/IDE/EditorTabs/EditableTab.tsx
@@ -1,12 +1,11 @@
-import React, { useMemo } from "react";
+import React, { useCallback } from "react";
 
-import { FileTab, type FileTabProps } from "IDE/Components/FileTab";
-import { useNameEditor } from "utils/hooks/useNameEditor";
+import { FileTab } from "IDE/Components/FileTab";
 import { type EntityItem } from "ee/entities/IDE/constants";
 import { useCurrentEditorState } from "../hooks";
 
 import { useSelector } from "react-redux";
-import { useEventCallback } from "usehooks-ts";
+import { useBoolean, useEventCallback } from "usehooks-ts";
 import { getIsSavingEntityName } from "ee/selectors/entitiesSelector";
 import { useFeatureFlag } from "utils/hooks/useFeatureFlag";
 import { FEATURE_FLAG } from "ee/entities/FeatureFlag";
@@ -14,11 +13,18 @@ import {
   getEditableTabPermissions,
   saveEntityName,
 } from "ee/entities/IDE/utils";
+import { noop } from "lodash";
+import { EditableName } from "IDE";
+import { IconContainer } from "IDE/Components/FileTab/styles";
 
-interface EditableTabProps extends Omit<FileTabProps, "isLoading" | "onClose"> {
+interface EditableTabProps {
   id: string;
+  onClick: () => void;
   onClose: (id: string) => void;
   entity?: EntityItem;
+  icon?: React.ReactNode;
+  isActive: boolean;
+  title: string;
 }
 
 export function EditableTab(props: EditableTabProps) {
@@ -31,23 +37,14 @@ export function EditableTab(props: EditableTabProps) {
     entity,
   });
 
-  const { handleNameSave, normalizeName, validateName } = useNameEditor({
-    entityId: id,
-    entityName: title,
-    nameSaveAction: (params) => saveEntityName({ params, segment, entity }),
-  });
+  const {
+    setFalse: exitEditMode,
+    setTrue: enterEditMode,
+    value: isEditing,
+  } = useBoolean(false);
 
   const isLoading = useSelector((state) =>
     getIsSavingEntityName(state, { id, segment, entity }),
-  );
-
-  const editorConfig = useMemo(
-    () => ({
-      onTitleSave: handleNameSave,
-      validateTitle: validateName,
-      titleTransformer: normalizeName,
-    }),
-    [handleNameSave, normalizeName, validateName],
   );
 
   const handleClose = useEventCallback((e: React.MouseEvent) => {
@@ -55,16 +52,32 @@ export function EditableTab(props: EditableTabProps) {
     onClose(id);
   });
 
+  const handleDoubleClick = isChangePermitted ? enterEditMode : noop;
+
+  const handleNameSave = useCallback(
+    (name: string) => {
+      saveEntityName({ params: { id, name }, segment, entity });
+      exitEditMode();
+    },
+    [entity, exitEditMode, id, segment],
+  );
+
   return (
     <FileTab
-      editorConfig={editorConfig}
-      icon={icon}
       isActive={isActive}
-      isChangePermitted={isChangePermitted}
-      isLoading={isLoading}
       onClick={onClick}
       onClose={handleClose}
+      onDoubleClick={handleDoubleClick}
       title={title}
-    />
+    >
+      <EditableName
+        exitEditing={exitEditMode}
+        icon={<IconContainer>{icon}</IconContainer>}
+        isEditing={isEditing}
+        isLoading={isLoading}
+        name={title}
+        onNameSave={handleNameSave}
+      />
+    </FileTab>
   );
 }
